corrected masking behavior for when binarization is disabled
Now, when binarize is disabled, script will determine the analysis mask area on the basis of counting occurrences of nonzero values at a given voxel across a sample. This is different than previously, where it would sum up the values at a given voxel across the sample and determine the analysis region on that basis. The "minimum lesion overlap" value is still used as a comparison. This was a problem because although the summing approach works when the data is binary, it would not work properly when the values were continuous. 

Also began introduction of analysis mask file (not enabled yet) - there is a new menu, but it is disabled.