added option for exclude disjoint lesions
In the past, a single lesion with no overlap with the rest of the lesion group would cause the subject to be excluded from an analysis. This behavior is now an option, with the default behavior being to exclude these disjoint lesions, as in the past. But this option can be disabled by Unchecking the option in the Image Data dropdown menu. Thanks for the request, Diego!